requirements.txt
fastapi
uvicorn
typing
tensorflow
pillow
numpy
opencv-python
azure-storage-blob